feature(Home): move styles and add animation
